feat: #23 不要な設定を削除した
