+++ web/src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from 'vue';
 import './styles/style.scss';
-import App from './pages/App.vue';
+import App from './App.vue';
 import router from './router/index';
 
 createApp(App).use(router).mount('#app');
