--- web/src/pages/App.vue
@@ -1,58 +1,0 @@
-<script setup lang="ts">
-import { onMounted } from 'vue';
-import HelloWorld from '../components/HelloWorld.vue';
-import axios, { AxiosResponse, AxiosError } from 'axios';
-
-const axiosInstance = axios.create({
-  baseURL: 'http://localhost:8080/api',
-  headers: {
-    'Content-Type': 'application/json',
-    'X-Requested-With': 'XMLHttpRequest'
-  },
-  responseType: 'json'
-});
-
-onMounted(() => {
-  axiosInstance
-    .get('/sample')
-    .then((response: AxiosResponse) => {
-      console.log(response);
-    })
-    .catch((err: AxiosError) => {
-      console.log(err);
-    });
-});
-</script>
-
-<template>
-  <div>
-    <a href="https://vitejs.dev" target="_blank">
-      <img src="../../public/vite.svg" class="logo" alt="Vite logo" />
-    </a>
-    <a href="https://vuejs.org/" target="_blank">
-      <img src="../assets/vue.svg" class="logo vue" alt="Vue logo" />
-    </a>
-    <a href="https://vuejs.org/" target="_blank">
-      <img src="../assets/sample.png" class="logo vue" alt="Vue logo" />
-    </a>
-  </div>
-  <router-link to="/">DashBoard</router-link>
-  <router-link to="/login">Login</router-link>
-  <router-view></router-view>
-  <HelloWorld msg="Vite + Vue" />
-</template>
-
-<style scoped>
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.vue:hover {
-  filter: drop-shadow(0 0 2em #42b883aa);
-}
-</style>
